start settup
added I belive all the folders and files needed for structure trough out the website.

added register and log in functionality with some simple styling to it.

added a feed page where all posts appear.

made it possible to create a post with title and content under with a simple post button.

added a nav bar with 3 simple ancors that navigates the UI to feed-page, profile-page and username button/w options for logout.
and a search bar for posts

added a profile page with follower count follow count and post count.

I also installed node to the project.
a update on the code, making it more readable will be done next.